+++ packages/widget/src/components/Messages/FundsSufficiencyMessage.tsx
@@ -1,9 +1,9 @@
 import { WarningRounded } from '@mui/icons-material'
-import { Typography } from '@mui/material'
+import { type BoxProps, Typography } from '@mui/material'
 import { useTranslation } from 'react-i18next'
-import { AlertMessage } from '../AlertMessage/AlertMessage.js'
+import { AlertMessage } from './AlertMessage.js'
 
-export const FundsSufficiencyMessage = () => {
+export const FundsSufficiencyMessage: React.FC<BoxProps> = (props) => {
   const { t } = useTranslation()
   return (
     <AlertMessage
@@ -13,7 +13,6 @@ export const FundsSufficiencyMessage = () => {
         <Typography
           variant="body2"
           sx={{
-            px: 1,
             color: 'text.primary',
           }}
         >
@@ -21,6 +20,7 @@ export const FundsSufficiencyMessage = () => {
         </Typography>
       }
       multiline
+      {...props}
     />
   )
 }