+++ packages/widget/src/components/Messages/GasRefuelMessage.tsx
@@ -6,8 +6,8 @@ import { useTranslation } from 'react-i18next'
 import { useGasRefuel } from '../../hooks/useGasRefuel.js'
 import { useSettings } from '../../stores/settings/useSettings.js'
 import { useSettingsActions } from '../../stores/settings/useSettingsActions.js'
-import { AlertMessage } from '../AlertMessage/AlertMessage.js'
-import { InfoMessageSwitch } from './GasMessage.style.js'
+import { AlertMessage } from './AlertMessage.js'
+import { InfoMessageSwitch } from './GasRefuelMessage.style.js'
 
 export const GasRefuelMessage: React.FC<BoxProps> = (props) => {
   const { t } = useTranslation()