+++ packages/widget/src/pages/SendToWallet/BookmarkAddressSheet.tsx
@@ -1,12 +1,16 @@
-import { Error as ErrorIcon, Info, TurnedIn } from '@mui/icons-material'
+import {
+  Error as ErrorIcon,
+  TurnedIn,
+  WarningRounded,
+} from '@mui/icons-material'
 import { Button, Typography } from '@mui/material'
 import type { ChangeEvent, MutableRefObject } from 'react'
 import { forwardRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { AlertMessage } from '../../components/AlertMessage/AlertMessage.js'
 import { BottomSheet } from '../../components/BottomSheet/BottomSheet.js'
 import type { BottomSheetBase } from '../../components/BottomSheet/types.js'
 import { Input } from '../../components/Input.js'
+import { AlertMessage } from '../../components/Messages/AlertMessage.js'
 import { useAddressValidation } from '../../hooks/useAddressValidation.js'
 import type { Bookmark } from '../../stores/bookmarks/types.js'
 import { useBookmarkActions } from '../../stores/bookmarks/useBookmarkActions.js'
@@ -206,11 +210,11 @@ export const BookmarkAddressSheet = forwardRef<
         </BookmarkInputFields>
         <AlertMessage
           title={
-            <Typography variant="body2">
-              {t('info.message.fundsToExchange')}
+            <Typography variant="body2" sx={{ color: 'text.primary' }}>
+              {t('warning.message.fundsLossPrevention')}
             </Typography>
           }
-          icon={<Info />}
+          icon={<WarningRounded />}
         />
         <SendToWalletButtonRow>
           <Button variant="text" onClick={handleCancel} fullWidth>