+++ packages/widget/src/pages/SendToWallet/ConfirmAddressSheet.tsx
@@ -1,12 +1,11 @@
-import { Info, Wallet, Warning } from '@mui/icons-material'
-import { Box, Button, Typography } from '@mui/material'
+import { Wallet, WarningRounded } from '@mui/icons-material'
+import { Button, Typography } from '@mui/material'
 import type { MutableRefObject } from 'react'
 import { forwardRef, useRef } from 'react'
 import { useTranslation } from 'react-i18next'
-import { AlertMessage } from '../../components/AlertMessage/AlertMessage.js'
 import { BottomSheet } from '../../components/BottomSheet/BottomSheet.js'
 import type { BottomSheetBase } from '../../components/BottomSheet/types.js'
-import { useIsContractAddress } from '../../hooks/useIsContractAddress.js'
+import { AlertMessage } from '../../components/Messages/AlertMessage.js'
 import { useNavigateBack } from '../../hooks/useNavigateBack.js'
 import { useSetContentHeight } from '../../hooks/useSetContentHeight.js'
 import type { Bookmark } from '../../stores/bookmarks/types.js'
@@ -23,7 +22,6 @@ import {
 interface ConfirmAddressSheetProps {
   onConfirm: (wallet: Bookmark) => void
   validatedBookmark?: Bookmark
-  chainId?: number
 }
 
 interface ConfirmAddressSheetContentProps extends ConfirmAddressSheetProps {
@@ -48,21 +46,15 @@ export const ConfirmAddressSheet = forwardRef<
 const ConfirmAddressSheetContent: React.FC<ConfirmAddressSheetContentProps> = ({
   validatedBookmark,
   onConfirm,
-  chainId,
   onClose,
 }) => {
   const { t } = useTranslation()
   const { navigateBack } = useNavigateBack()
   const { setFieldValue } = useFieldActions()
   const { setSendToWallet } = useSendToWalletActions()
-  const isContractAddress = useIsContractAddress(
-    validatedBookmark?.address,
-    chainId,
-    validatedBookmark?.chainType
-  )
 
   const containerRef = useRef<HTMLElement>(null)
-  useSetContentHeight(containerRef, isContractAddress)
+  useSetContentHeight(containerRef)
 
   const handleConfirm = () => {
     if (validatedBookmark) {
@@ -97,28 +89,15 @@ const ConfirmAddressSheetContent: React.FC<ConfirmAddressSheetContentProps> = ({
         <Typography>{validatedBookmark?.address}</Typography>
       </SheetAddressContainer>
       <AlertMessage
+        severity="warning"
         title={
-          <Box sx={{ display: 'grid', gap: 1 }}>
-            <Typography variant="body2" fontWeight={500}>
-              {t('info.message.fundsToExchange')}
-            </Typography>
-          </Box>
+          <Typography variant="body2" sx={{ color: 'text.primary' }}>
+            {t('warning.message.fundsLossPrevention')}
+          </Typography>
         }
-        icon={<Info />}
+        icon={<WarningRounded />}
         multiline
       />
-      {isContractAddress ? (
-        <AlertMessage
-          title={
-            <Typography variant="body2" fontWeight={500}>
-              {t('info.message.smartContractAccount')}
-            </Typography>
-          }
-          icon={<Warning />}
-          severity="warning"
-          multiline
-        />
-      ) : null}
       <SendToWalletButtonRow>
         <Button variant="text" onClick={onClose} fullWidth>
           {t('button.cancel')}